backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
python-multipart==0.0.17
anthropic>=0.49.0
